Refactor global styles and components: consolidate LayoutRow and TabSwitcher styles into globals, update imports across components and pages for consistency.
