+++ royal-care-frontend/README.md
@@ -47,7 +47,8 @@ royal-care-frontend/
 │   │   ├── MainLayout.css              # Layout styling
 │   │   ├── CompanyInfo.css             # Company information styling
 │   │   ├── Placeholders.css            # Placeholder content styling
-│   │   └── theme.css                   # Global theme variables
+│   │   ├── theme.css                   # Global theme variables (USE THESE!)
+│   │   └── app.css                     # Global app styles (USE THESE!)
 │   ├── App.jsx                         # Main app component and routing
 │   ├── main.jsx                        # Entry point
 │   └── store.js                        # Redux store configuration
@@ -141,12 +142,58 @@ This frontend connects to the Django backend located in the `/guitara` directory
 - [Supabase Documentation](https://supabase.com/docs)
 - [React Router Documentation](https://reactrouter.com/)
 
-## Contributing
+## Frontend Coding Rules & Guidelines
 
-1. Follow the project's coding standards and file organization
-2. Update CSS in the appropriate files in the `src/styles` directory
-3. Add new components in the appropriate directories based on their function
-4. Test features across different roles before submitting changes
+> **Please follow these rules for all frontend development!**
+
+1. **Use Global Variables:**
+
+   - All color, spacing, and font variables are defined in `theme.css` and `app.css`. **Always use these variables!**
+   - Do not hardcode colors, spacing, or font sizes unless absolutely necessary.
+
+2. **Use Global Components:**
+
+   - For tabs, use `TabSwitcher` and its styles in `tabswitcher.css`.
+   - For layout rows, use `LayoutRow` (`layoutrow.jsx`) and its styles in `layoutrow.css`.
+   - **Do not create redundant tab or row components.**
+
+3. **Consistency:**
+
+   - Reuse global styles and components for a consistent look and feel.
+   - Avoid duplicating logic or UI patterns that already exist in the codebase.
+
+4. **Box Shadows:**
+
+   - **Do NOT use `box-shadow`.**
+   - If you need to blur or dim the background (e.g., for modals or overlays), use:
+     ```css
+     background: rgba(0, 0, 0, 0.6);
+     backdrop-filter: blur(4px);
+     -webkit-backdrop-filter: blur(4px);
+     ```
+   - This applies to overlays, modals, and any feature requiring background dimming or blur.
+
+5. **CSS Location:**
+
+   - Place all new CSS in the appropriate file in `src/styles/`.
+   - Do not inline styles unless absolutely necessary.
+
+6. **Component Placement:**
+
+   - Add new components in the correct directory based on their function.
+   - Avoid creating duplicate or redundant components.
+
+7. **Testing:**
+   - Test all features across different user roles and screen sizes before submitting changes.
+
+## Pages Folder Structure
+
+- The `src/pages/` directory contains a separate folder for each page in the application.
+- **Example:** For the Settings Data page, there is a folder named `SettingsDataPage` (case-sensitive, match existing convention), and inside that folder are the main `.jsx` file and a corresponding `.module.css` file for styles.
+- **When creating new pages:**
+  - Always create a new folder inside `src/pages/` for the page.
+  - Place the main component (e.g., `MyNewPage.jsx`) and its CSS module (e.g., `MyNewPage.module.css`) inside that folder.
+  - This keeps code organized and makes it easy to manage page-specific logic and styles.
 
 ---
