+++ royal-care-frontend/src/globals/TabSwitcher.css
