+++ royal-care-frontend/src/pages/SettingsDataPage/SettingsDataPage.module.css
@@ -1,4 +1,4 @@
-@import "../../styles/TabSwitcher.css";
+@import "../../globals/TabSwitcher.css";
 
 /* Place your SettingsDataPage specific styles here */
