--- royal-care-frontend/src/styles/theme.css
@@ -1,175 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Plus%20Jakarta%20Sans:700|Plus%20Jakarta%20Sans:400"); /* Import Google Fonts */
-
-/* Base HTML styles */
-html {
-  font-size: 100%; /* 16px */
-  color: var(--text);
-  background: var(--background);
-}
-
-/* Body styles */
-body {
-  color: var(--text);
-  background: var(--background);
-}
-
-/* Set font family and weight for common elements */
-body,
-input,
-button,
-textarea,
-select,
-label,
-a,
-p,
-div,
-span {
-  font-family: "Plus Jakarta Sans", sans-serif !important;
-  font-weight: 400;
-}
-
-/* Headings use bold font weight */
-h1,
-h2,
-h3,
-h4,
-h5 {
-  font-family: "Plus Jakarta Sans", sans-serif !important;
-  font-weight: 700;
-}
-
-/* Heading font sizes using CSS variables */
-h1 {
-  font-size: var(--font-size-xxxl);
-}
-h2 {
-  font-size: var(--font-size-xxl);
-}
-h3 {
-  font-size: var(--font-size-xl);
-}
-h4 {
-  font-size: var(--font-size-lg);
-}
-h5 {
-  font-size: var(--font-size-md);
-}
-
-/* Small text size */
-small {
-  font-size: var(--font-size-sm);
-}
-
-/* Theme color and spacing variables */
-:root {
-  /* Main colors */
-  --text: #0f0c05;
-  --muted: var(--background-500);
-  --background: #ffffff;
-  --primary: #4b3b06;
-  --primary-hover: var(--primary-900);
-  --secondary: #adbdae;
-  --accent: #9ab096;
-  --error: #ff312e;
-
-  /* Text color shades */
-  --text-50: #f9f5ec;
-  --text-100: #f2ebd9;
-  --text-200: #e6d6b3;
-  --text-300: #d9c28c;
-  --text-400: #ccad66;
-  --text-500: #bf9940;
-  --text-600: #997a33;
-  --text-700: #735c26;
-  --text-800: #4d3d19;
-  --text-900: #261f0d;
-  --text-950: #130f06;
-
-  /* Background color shades */
-  --background-25: #fafafa; /* lighter than background-50 */
-  --background-35: #f8f8f8; /* between 25 and 40, a bit lighter */
-  --background-40: #f6f6f6; /* between 25 and 50 */
-  --background-50: #f2f2f2;
-  --background-100: #e6e6e6;
-  --background-200: #cccccc;
-  --background-300: #b3b3b3;
-  --background-400: #999999;
-  --background-500: #808080;
-  --background-600: #666666;
-  --background-700: #4d4d4d;
-  --background-800: #333333;
-  --background-900: #1a1a1a;
-  --background-950: #0d0d0d;
-
-  /* Primary color shades */
-  --primary-50: #fdf8e7;
-  --primary-100: #fbf1d0;
-  --primary-200: #f7e3a1;
-  --primary-300: #f4d571;
-  --primary-400: #f0c742;
-  --primary-500: #ecb913;
-  --primary-600: #bd940f;
-  --primary-700: #8e6f0b;
-  --primary-800: #5e4a08;
-  --primary-900: #2f2504;
-  --primary-950: #181302;
-
-  /* Secondary color shades */
-  --secondary-50: #f1f4f1;
-  --secondary-100: #e3e8e3;
-  --secondary-200: #c6d2c7;
-  --secondary-300: #aabbab;
-  --secondary-400: #8ea48f;
-  --secondary-500: #718e73;
-  --secondary-600: #5b715c;
-  --secondary-700: #445545;
-  --secondary-800: #2d392e;
-  --secondary-900: #171c17;
-  --secondary-950: #0b0e0c;
-
-  /* Accent color shades */
-  --accent-50: #f1f4f0;
-  --accent-100: #e3e9e2;
-  --accent-200: #c7d3c5;
-  --accent-300: #abbda8;
-  --accent-400: #8fa78b;
-  --accent-500: #73916e;
-  --accent-600: #5c7458;
-  --accent-700: #455742;
-  --accent-800: #2e3a2c;
-  --accent-900: #171d16;
-  --accent-950: #0c0f0b;
-
-  /* Status colors */
-  --success-bg: #e6f4ea;
-  --success-text: #137333;
-  --error-bg: #fce8e6;
-  --error-text: #d93025;
-
-  /* Font sizes */
-  --font-size-xs: 0.7rem;
-  --font-size-sm: 0.8rem;
-  --font-size-base: 1rem;
-  --font-size-md: 1.25rem;
-  --font-size-lg: 1.563rem;
-  --font-size-xl: 1.954rem;
-  --font-size-xxl: 2.442rem;
-  --font-size-xxxl: 3.053rem;
-
-  /* Spacing variables */
-  --spacing-xxs: 0.25rem; /* 4px */
-  --spacing-xs: 0.5rem; /* 8px */
-  --spacing-xm: 0.75rem; /* 12px */
-  --spacing-sm: 1rem; /* 16px */
-  --spacing-mm: 1.5rem; /* 24px */
-  --spacing-mmd: 1.75rem; /* 28px */
-  --spacing-md: 2rem; /* 32px */
-  --spacing-ml: 2.5rem; /* 40px */
-  --spacing-lg: 3rem; /* 48px */
-  --spacing-xl: 4rem; /* 64px */
-  --spacing-xxl: 5rem; /* 80px */
-  --spacing-xxxl: 6rem; /* 96px */
-
-  /* Container max width: global variable for the width of the main content container on all pages. Adjust this to control the maximum width of page content for consistent layouts. */
-  --container-max-width: 1000px;
-}
